Add files via upload
I recently completed a graded SQL assignment as part of the IBM Data Science Professional Certificate on Coursera. The assignment covered various aspects of SQL, including querying databases, managing data, and using SQL for data analysis. Throughout the course, I gained a solid understanding of how to retrieve and manipulate data using SQL commands, as well as the importance of database design and optimization. The hands-on nature of the assignment allowed me to apply theoretical concepts to practical scenarios, reinforcing my skills in working with databases. Completing this assignment was a valuable experience in honing my SQL skills, and I'm excited to continue applying these learnings in real-world data science projects. The IBM Data Science Professional Certificate provides a comprehensive and practical foundation for anyone looking to excel in the field of data science.